Bug 1434695 - Load cached clients when showing the client picker in Send Tab.
Previously we would kick off a refresh spinner, do a full sync of
clients, then fetch them from the DB.

Most users' client lists will almost never change.

With this PR, we first fetch the client list from the DB. If there are
no clients, we kick off a regular refresh with spinner. If there are
clients, we silently run a client sync (no spinner) and update the
displayed list if it changes.

Pull-to-refresh is unaffected.